data-mgmt add api

+++ yk-common/yk-common-utils/src/main/java/org/yk/common/data/PageInfo.java
@@ -29,5 +29,5 @@ public class PageInfo implements Serializable {
 	/**
 	 * Optional. totalRecords: total records of the entities.
 	 */
-	private Integer totalRecords;
+	private Long totalRecords;
 }
